Update: Created sub-folder for welcome-page & Renamed components folder
