+++ frontend/src/components/welcome-page/WelcomePage.css
