+++ frontend/src/components/welcome-page/index.js
